% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cbp_data.R
\name{download_all_cbp}
\alias{download_all_cbp}
\title{Download the CBP}
\usage{
download_all_cbp(year_start = 1996, year_end = 2013,
  aggregation_level = "county", path_data = "~/Downloads/")
}
\arguments{
\item{year_start:}{year starting data download}

\item{year_end:}{year ending data download}

\item{aggregation_level:}{which data type to download}

\item{path_data:}{where does the download happen: default current directory}
}
\value{
data.table
}
\description{
Download the CBP
}
\note{
downloads data from the County Business Pattern
  for desired years, splitting data by the aggregation level codes
  the docs for the CBP is available at the following:
  http://www.census.gov/econ/cbp/download/full_layout/County_Layout_SIC.txt
}
