% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qcew_data.R
\name{download_qcew_sic_data}
\alias{download_qcew_sic_data}
\title{Download QCEW dataset from directly from the BLS website for SIC Industries}
\usage{
download_qcew_sic_data(target_year, path_data = "./",
  frequency = "quarter", url_wayback = "", unzip_dir = T,
  download = T, verbose = F)
}
\arguments{
\item{target_year}{year for which we want to download the data}

\item{path_data}{where does the download happen: default current directory + tmp}

\item{frequency}{download the quarterly files or the yearly files (default is quarterly)}

\item{url_wayback}{allows to specify the path in internet wayback machine that kept some of the archive}

\item{unzip_dir}{unzip the file}

\item{download}{default is TRUE, FALSE only returns the url}

\item{verbose}{how does the function output intermediate stages}
}
\value{
file complete path. Downloads the file to the current directory and unzips it.
}
\description{
Download QCEW dataset from directly from the BLS website for SIC Industries
}
