% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qcew_data.R
\name{download_qcew_size_data}
\alias{download_qcew_size_data}
\title{Download QCEW dataset (size files 1st quarter) directly from the BLS website}
\usage{
download_qcew_size_data(target_year, industry = "naics", path_data = "./")
}
\arguments{
\item{target_year:}{year for which we want to download the data}

\item{path_data:}{where does the download happen: default current directory}

\item{industry:}{download naics or sic data}
}
\value{
NIL. Downloads the file to the current directory and unzips it.
}
\description{
Download QCEW dataset (size files 1st quarter) directly from the BLS website
}

