% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qcew_data.R
\name{export_all_data}
\alias{export_all_data}
\title{Export a dataset we are directly reading.}
\usage{
export_all_data(target_year, path_data = "./")
}
\arguments{
\item{path_data:}{where does the download happen: default current directory}

\item{target_year:}{year for which we want to export the data}
}
\value{
NIL. Exports the data using the helper function export_named_df.
}
\description{
Export a dataset we are directly reading.
}
\note{
assumes that file is at "./$target_year.q1_q4.singlefile.csv". This is the expected
  path and file name when the csv file is extracted from the zip file with no renaming.
}
