% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qcew_data.R
\name{export_named_df}
\alias{export_named_df}
\title{Export the downloaded table into a csv file.}
\usage{
export_named_df(path_data = "./", year, named_df)
}
\arguments{
\item{year:}{filename for .csv file. Note that you need to pass in the
correct year here, since the function has no idea of what the actual year
corresponding to this data is}

\item{named_df:}{data frame with name}

\item{path_data:}{where does the download happen: default current directory}
}
\value{
NIL. Exports csv file "./singlefile/$agglvl_code/$year.csv"
}
\description{
Export the downloaded table into a csv file.
}
\note{
generates the appropriate directory if it does not already exist.

our data does not have any headers; this simplifies the process of
  joining all the csv files with cat or similar command line tools so that we
  have a single file across all aggregation level codes.
}
