% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qcew_data.R
\name{fread.zip.url}
\alias{fread.zip.url}
\title{Other code: read zipped url}
\usage{
fread.zip.url(url, filename = NULL, FUN = fread, ...)
}
\arguments{
\item{url}{see vignette for a list of the cut}

\item{filename}{for multiple files in the archive}

\item{FUN}{default to fread from data.table but you could use anything}

\item{...}{arguments passed to the function call (typically fread)}
}
\value{
data.table of the remote file
}
\description{
Other code: read zipped url
}
\note{
This code is directly copied from https://stackoverflow.com/a/24586478
}
\examples{
  \dontrun{
  dt <- fread.zip.url(http://data.bls.gov/cew/data/files/2000/csv/2000_qtrly_singlefile.zip)
  }
}
