% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bds_data.R
\name{get_bds_cut}
\alias{get_bds_cut}
\title{Other code: loads data.table with option to write it for a given cut}
\usage{
get_bds_cut(year_start = 1977, year_end = 2014, unit = "firm",
  aggregation = "all")
}
\arguments{
\item{year_start}{start year for which we want data}

\item{year_end}{end year for which we want data}

\item{unit}{establishment level or firm level}

\item{aggregation}{download the quarterly files or the yearly files (default is quarterly)}
}
\value{
data.table aggregate
}
\description{
Other code: loads data.table with option to write it for a given cut
}
\examples{
\dontrun{
  dt <- get_bds_cut(year_start = 1980, year_end = 2012,
                    unit = "firm", aggregation = "all")
}
}
