% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bed_data.R
\name{get_bed}
\alias{get_bed}
\title{Group all the functions}
\usage{
get_bed(which_data = "industry", level = FALSE, path_data = NULL)
}
\arguments{
\item{which_data}{what kind of data download: default is the main industry file}

\item{level}{also keep the level of the variables (number of estabs and employment counts)}

\item{path_data}{where does the download happen: default current directory (irrelevant since we read mostly from url)}
}
\value{
df
}
\description{
Group all the functions
}
\note{
downloads data from the Entry Data from the Business Employment Dynamics Website
  Data is split and tidy
Download BED dataset from directly from the BLS website
}
