% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bed_data.R
\name{get_bed_data}
\alias{get_bed_data}
\title{Group all the functions}
\usage{
get_bed_data(which_data = "industry", path_data = "./")
}
\arguments{
\item{which_data:}{what kind of data download: default is the main industry file}

\item{path_data:}{where does the download happen: default current directory}
}
\value{
df_res.
}
\description{
Group all the functions
}
\note{
downloads data from the Entry Data from the Business Employment Dynamics Website
  Data is split and tidy
Download BED dataset from directly from the BLS website
}
