% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bed_data.R
\name{get_bed_detail}
\alias{get_bed_detail}
\title{Download BED dataset from directly from the BLS website}
\usage{
get_bed_detail(seasonal = "S", msa = "00000", state = "00",
  county = "000", industry = "000000", unitanalysis = 1,
  dataelement = 2, sizeclass = "00", dataclass = c("07"),
  ratelevel = "L", periodicity = "Q", ownership = 5, download = F,
  path_data = "./")
}
\arguments{
\item{seasonal:}{seasonality of the data; S (default) or U}

\item{msa:}{msa code; 00000 (default) for national}

\item{state:}{state code; 00 (default) for national}

\item{county:}{county code; 000 (default) for national}

\item{industry:}{industry code; 000000 (default) for total private. See docs for industry classifications}

\item{unitanalysis:}{unit of observation; 1 (default) for establishment}

\item{dataelement:}{which type of date; 2 (default) number of establishments, 1 employment}

\item{sizeclass:}{firm size class; 00 (default) for all}

\item{dataclass:}{for type of information; 07 (default) for Establishment Births, 01 Gross Job Gains, 02 Expansions, 03 Openings, 04 Gross Job Losses, 05 Contractions, 06 Closings, 07 Establishment Births, 08 Establishment Deaths}

\item{ratelevel:}{for type of statistics; L (default) for level, R for rate}

\item{periodicity:}{for sampling frequency; Q (default) for quarterly, A for annual}

\item{ownership:}{for type of ownership; 5 (default) for private sector}

\item{download:}{requires download confirmation}

\item{path_data:}{where does the download happen: default current directory}
}
\value{
dt_final
}
\description{
Download BED dataset from directly from the BLS website
}
