% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qcew_data.R
\name{get_files_cut}
\alias{get_files_cut}
\title{Other code: loads data.table with option to write it for a given cut}
\usage{
get_files_cut(data_cut = 10, year_start = 1990, year_end = 2013,
  path_data = "~/Downloads/", write = F)
}
\arguments{
\item{path_data:}{where do we store the data}

\item{year_start:}{start year for which we want data}

\item{year_end:}{end year for which we want data}
}
\value{
data.table aggregate
}
\description{
Other code: loads data.table with option to write it for a given cut
}
\examples{
  dt <- get_files_cut(data_cut = 10, year_start = 1990, year_end =1993, path_data = "~/Downloads/", write = T)
}

