% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lau.R
\name{get_lau_data}
\alias{get_lau_data}
\title{Group all the functions}
\usage{
get_lau_data(which_data = "county", years = seq(1990, 2016),
  path_data = "./", verbose = T)
}
\arguments{
\item{which_data}{what kind of data download: default are the county files}

\item{years}{which year do we download, as of now only 1990 to 2016 are available}

\item{path_data}{where does the download happen: default current directory}

\item{verbose}{show intermediate steps}
}
\value{
dt_res
}
\description{
Group all the functions
}
\note{
downloads data from the Local Area Unemployment Statistics from the BLS
#'   Data is split and tidy
Download LAU County dataset from directly from the BLS website
}
\examples{
  \dontrun{
    dt_lau <- get_lau_data(years = c(1995, 2000))
  }
}
