% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nbf.R
\name{get_nbf}
\alias{get_nbf}
\title{Other code: loads data.table with option to write it for a given cut}
\usage{
get_nbf(year_start = 1948, year_end = 2014)
}
\arguments{
\item{year_start}{start year for which we want data}

\item{year_end}{end year for which we want data}
}
\value{
data.table aggregate
}
\description{
Other code: loads data.table with option to write it for a given cut
}
\examples{
  \dontrun{
  dt <- get_nbf(year_start = 1980, year_end = 2012)
 }
}
