% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scb.R
\docType{data}
\name{scb}
\alias{scb}
\title{Survey of Current Business Extracted File}
\format{Data Table (data.table) with columns
\describe{
\item{date_ym}{Data}
\item{nbf_annual}{Net business Formation (annual)}
\item{nbf}{Net business Formation (monthly)}
\item{nbi_annual}{New business Incorporation (annual)}
\item{nbi}{New business Incorporation (monthly)}
\item{cbf_annual}{New business Incorporation (annual)}
\item{cbf}{New business Incorporation (monthly)}
}}
\source{
https://www.bea.gov/scb/pdf/NATIONAL/BUSCYCLE/1996/0296cpgs.pdf
}
\usage{
scb
}
\description{
Survey of Current Business Extracted File
}
\keyword{datasets}
