% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cbp_data.R
\name{tidy_cbp}
\alias{tidy_cbp}
\title{Clean the CBP}
\usage{
tidy_cbp(dt_raw = NULL, industry = c(2, 3, 4), year_start = 1996,
  year_end = 2013)
}
\arguments{
\item{industry:}{which level industries should be included (vector, NULL for all)}

\item{year_start:}{year starting data (>=1986)}

\item{year_end:}{year ending data (<=2013)}
}
\value{
data.table
}
\description{
Clean the CBP
}
\note{
Tidy up the CBP data into year state county industry dataset
}

