% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qcew_data.R
\name{tidy_qcew_year}
\alias{tidy_qcew_year}
\title{Tidying the dataset for regression use (or merge): operate year by year to keep memory to manageable levels.}
\usage{
tidy_qcew_year(dt, frequency = "all", industry = "naics")
}
\arguments{
\item{dt:}{input dataset from get_files_cut}

\item{industry:}{download naics or sic data}

\item{frequency:}{download either quarterly, monthly or all data}
}
\value{
data.table dt_res
}
\description{
Tidying the dataset for regression use (or merge): operate year by year to keep memory to manageable levels.
}
\note{
returns a data.table file that is formatted according to tidy standard
  typically this will be year x sub_year (quarter or month) x size x own_code x industry
  the file can be aggregated as such
  I do not download all the information (some location quotients and taxes are forgotten)
}
\examples{
  dt_tidy <- tidy_qcew_year(dt, frequency = "all", industry = "naics")
}
