% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cbp_data.R
\name{download_cbp_data}
\alias{download_cbp_data}
\title{Download one year of county business pattern dataset}
\usage{
download_cbp_data(target_year, aggregation_level = "county",
  path_data = "~/Downloads/")
}
\arguments{
\item{target_year:}{year for which we want to download the data}

\item{aggregation_level:}{which data cut to download}

\item{path_data:}{where does the download happen: default current directory}
}
\value{
data.table
}
\description{
Download one year of county business pattern dataset
}
