% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qcew_data.R
\name{download_qcew_data}
\alias{download_qcew_data}
\title{Group all the functions}
\usage{
download_qcew_data(target_year, path_data = "./")
}
\arguments{
\item{target_year:}{year for which we want to download the data}

\item{path_data:}{where does the download happen: default current directory}
}
\value{
NIL. Downloads the file to the current directory and unzips it.
}
\description{
Group all the functions
}
\note{
downloads data from the Quarterly Census for Employment and Wages
  program, for desired years, splitting data by the aggregation level codes
  and exporting them into separate files sorted by year and aggregation level

: we intend to have a single file for each aggregation level, but it is
more convenient to merge the individual files using an external command line
tool like cat

Download QCEW dataset from directly from the BLS website
}

