% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qcew_data.R
\name{download_qcew_size_data}
\alias{download_qcew_size_data}
\title{Download QCEW dataset (size files 1st quarter) directly from the BLS website}
\usage{
download_qcew_size_data(target_year, industry = "naics", path_data = "./",
  url_wayback = "", unzip = F)
}
\arguments{
\item{target_year:}{year for which we want to download the data}

\item{path_data:}{where does the download happen: default current directory}

\item{industry:}{download naics or sic data}

\item{url_wayback:}{allows to specify the path in internet wayback machine that kept some of the archive}

\item{unzip:}{default is True, False is useful if you are just downloading the data}
}
\value{
read_list: String with names of downloaded files
}
\description{
Download QCEW dataset (size files 1st quarter) directly from the BLS website
}
\note{
Downloads the file to the current directory and unzips it.
}
